Use uint32_t for key IDs and trie/node counts in the public API
This is a breaking change for the API and ABI since it affects return
types, arguments, and class field sizes.

Existing source code should still build, but if -Wconversion is enabled,
it may warn about narrowing conversions for key IDs on 64-bit systems.

Dynamically linked programs must be rebuilt on 64-bit systems.

Note that these have always been at most a uint32_t internally.

+++ include/marisa/base.h
@@ -29,7 +29,7 @@ using marisa_uint64 [[deprecated]] = std::uint64_t;
 [[deprecated]] constexpr auto MARISA_SIZE_MAX = SIZE_MAX;
 
 // Key IDs are always 32-bit unsigned integers.
-using marisa_key_t = size_t;
+using marisa_key_t = uint32_t;
 
 #define MARISA_INVALID_LINK_ID UINT32_MAX
 #define MARISA_INVALID_KEY_ID  UINT32_MAX
